Refactor endpoint
- Change URL to `github-token`.
- Move endpoint code to dedicated class.
